Having spec & resource share their registration code adds ability to have multiple values for spec attributes like resource supports.
BaseNFSDiskSpec's override of BaseDiskSpec was a somewhat tricky problem, as it ran afoul of validation on setting all attributes (something that makes sense for Resource's `REQUIRED_ATTR` name). Removed the "no None attributes" check from registration, but Get should still catch it. Spec now also uses resource's Get, but allows for a not found subclass to fall back to the base class.

PiperOrigin-RevId: 601287749