Add a new required attribute (PLATFORM) to VirtualMachine allowing for more variety
Technically many of the implementations of VirtualMachine are not virtual machines, but something which still needs to run commands & install packages. The PLATFORM attributes allows for those implementations to specify their type more accurately, and without the overload of CLOUD which they currently use.

In follow-ups this will allow for e.g. Kubernetes & Static to not use CLOUD=Kubernetes but rather VM_PLATFORM=Kubernetes & CLOUD=GCP or AWS as expected.

PiperOrigin-RevId: 605669678